Reading and parsing files from folders generatedMessages and referenceMessages

+++ compare_xml_using_xmldiff.py
@@ -51,7 +51,7 @@ def get_type(text):
 
 
 # Custom message map
-def custom_message(op, inserted_nodes, deleted_nodes, ignored_tags):
+def custom_message(op, inserted_nodes, deleted_nodes, ignored_tags, tree1):
     if isinstance(op, actions.DeleteNode):
         parent_path = '/'.join(op.node.split('/')[0:-1])
 
@@ -100,10 +100,38 @@ def load_config(path):
         data = json.load(file)
         return data
 
+def run_comparaison(reference_path:str, generated_path:str, config_path:str):
+    # Parse XML files
+    tree1 = etree.parse(file1)
+    tree2 = etree.parse(file2)
+
+    # Run xmldiff in 'diff_files' mode to get structured actions
+    ops = main.diff_files(reference_path, generated_path, diff_options={'F': 0.1})  # F=similarity threshold
+
+    # ops.sort(key=lambda x: 0 if isinstance(x, actions.DeleteNode) or isinstance(x, actions.InsertNode) else 1)
+
+    config = load_config(config_path)
+    deleted_nodes = find_deletes(ops)
+    inserted_nodes = set()
+    ignored_tags = config["ignored_tags"]
+    # ignored_tags = ["UUID", "TS", "SessionID"]
+    # deleted_nodes = set()
+    # Display custom messages
+    for op in ops:
+        msg = custom_message(op, inserted_nodes, deleted_nodes, ignored_tags, tree1)
+        if msg:
+            print(msg)
+
+
 for test in tests:
     print("---------------------------------")
     print("Running test case:")
     print(test["label"])
+
+    if (test["label"] == "9a- Existence + valeur unique" or  test["label"] == "9c- Valeur + exclusion de champ"
+            or  test["label"] == "10b – Exclusion de tag commentaire variable"):
+        continue
+
     #file1 = f"reference-{test["id"]}.xml"
     file1 = f"reference.xml"
     with open(file1, "w", encoding="utf-8") as file:
@@ -113,24 +141,4 @@ for test in tests:
     with open(file2, "w", encoding="utf-8") as file:
         file.write(test["xml_generated"])
 
-    # Parse XML files
-    tree1 = etree.parse(file1)
-    tree2 = etree.parse(file2)
-
-    # Run xmldiff in 'diff_files' mode to get structured actions
-    ops = main.diff_files(file1, file2, diff_options={'F': 0.1})  # F=similarity threshold
-
-    # ops.sort(key=lambda x: 0 if isinstance(x, actions.DeleteNode) or isinstance(x, actions.InsertNode) else 1)
-
-
-    config = load_config("config.json")
-    deleted_nodes = find_deletes(ops)
-    inserted_nodes = set()
-    ignored_tags = config["ignored_tags"]
-    #ignored_tags = ["UUID", "TS", "SessionID"]
-    # deleted_nodes = set()
-    # Display custom messages
-    for op in ops:
-        msg = custom_message(op, inserted_nodes, deleted_nodes, ignored_tags)
-        if msg:
-            print(msg)
+    run_comparaison(file1,file2, "config.json")
